FlaskApi/requirements.txt
aniso8601==3.0.2
astroid==2.0
atomicwrites==1.1.5
attrs==18.1.0
autoenv==1.0.0
beautifulscraper==1.1.0
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.1
django-crispy-forms==1.7.2
djangorestframework==3.7.7
flake8==3.5.0
Flask==1.0.2
Flask-API==1.0
Flask-Cors==3.0.6
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
flask-pytest==0.0.5
flask-restplus==0.11.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gunicorn==19.8.1
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.2.0
mysqlclient==1.3.12
nose==1.3.7
passlib==1.7.1
pbr==4.1.0
Pillow==5.1.0
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pylint==2.0.0
PyMySQL==0.8.1
pytest==3.6.3
pytz==2017.3
requests==2.19.1
reverse==0.1.0
six==1.11.0
SQLAlchemy==1.2.7
stevedore==1.28.0
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
